requirements.txt
aiogram==3.4.1
Pillow==10.4.0
python-dotenv==1.0.1
aiosqlite==0.19.0
cachetools==5.3.2
# Web application
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.2
itsdangerous==2.1.2